Add coverage for MkDocs scaffolder
